refactor: 코드 구조 변경
- repository 역할에 따라서 분리
- Service 역할에 따라 분리
- Usecase가 많아질 것을 대비하여 Facade 추가
