+++ backend/queue-service/src/main/java/com/bidket/queue/application/service/QueueTrafficService.java
@@ -1,14 +1,11 @@
 package com.bidket.queue.application.service;
 
 import com.bidket.queue.domain.exception.QueueException;
-import com.bidket.queue.domain.model.QueueConfigModel;
 import com.bidket.queue.domain.model.QueueErrorCode;
 import com.bidket.queue.domain.model.QueueStatus;
-import com.bidket.queue.domain.repository.RedisRepository;
+import com.bidket.queue.domain.repository.QueueTrafficRepository;
 import com.bidket.queue.global.annotation.CheckQueueConfig;
 import com.bidket.queue.global.util.jwt.TokenProvider;
-import com.bidket.queue.presentation.dto.request.QueueCreateRequest;
-import com.bidket.queue.presentation.dto.response.QueueCreateResponse;
 import com.bidket.queue.presentation.dto.response.QueueEnterResponse;
 import com.bidket.queue.presentation.dto.response.QueueStatusResponse;
 import lombok.RequiredArgsConstructor;
@@ -16,55 +13,21 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
 
-import java.time.temporal.ChronoUnit;
 import java.util.UUID;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class QueueService {
-    private final RedisRepository redisRepository;
+public class QueueTrafficService {
+    private final QueueTrafficRepository trafficRepository;
     private final TokenProvider tokenProvider;
-
-    public Mono<QueueCreateResponse> createConfigQueue(QueueCreateRequest request) {
-        String key = "queue:auction:" + request.auctionId() + ":config";
-        QueueConfigModel queueConfig = request.toModel();
-        return redisRepository.saveConfig(key, queueConfig)
-                .flatMap(isSuccess -> {
-                    if (!isSuccess)
-                        return Mono.error(new QueueException(QueueErrorCode.REDIS_SAVE_FAILED));
-
-                    return redisRepository.setExpiration(key, request.closeAt().plus(1, ChronoUnit.DAYS));
-                })
-                .onErrorResume(e -> {
-                    if (e instanceof QueueException)
-                        return Mono.error(e);
-
-                    return redisRepository.deleteConfig(key)
-                            .then(Mono.error(new QueueException(QueueErrorCode.REDIS_EXPIRE_SET_FAILED)));
-                })
-                .flatMap(isExpireSuccess -> {
-                    if (isExpireSuccess) {
-                        return redisRepository.registerActiveAuction(request.auctionId())
-                                .onErrorMap(e -> new QueueException(QueueErrorCode.AUCTION_REGISTER_FAIL))
-                                .map(addedCount -> queueConfig.toCreateResponse());
-                    }
-                    return Mono.error(new QueueException(QueueErrorCode.REDIS_EXPIRE_SET_FAILED));
-                })
-                .onErrorMap(e -> {
-                    if (e instanceof QueueException)
-                        return e;
-
-                    return new QueueException(QueueErrorCode.REDIS_CONNECTION_ERROR);
-                });
-    }
 
     @CheckQueueConfig
     public Mono<QueueEnterResponse> enterQueue(UUID userId, UUID auctionId) {
         String waitingKey = "queue:auction:" + auctionId + ":waiting";
 
-        return redisRepository.addWaitingUser(waitingKey, userId)
-                .flatMap(isAdded -> redisRepository.getRank(waitingKey, userId))
+        return trafficRepository.addWaitingUser(waitingKey, userId)
+                .flatMap(isAdded -> trafficRepository.getRank(waitingKey, userId))
                 .map(rank -> QueueEnterResponse.builder()
                         .auctionId(auctionId)
                         .userId(userId)
@@ -79,7 +42,7 @@ public class QueueService {
         String tokenKey = "queue:token:" + auctionId;
         String waitingKey = "queue:auction:" + auctionId + ":waiting";
 
-        return redisRepository.getToken(tokenKey, userId)
+        return trafficRepository.getToken(tokenKey, userId)
                 .flatMap(token -> {
                     if (!tokenProvider.validateToken(token, userId, auctionId))
                         return Mono.error(new QueueException(QueueErrorCode.INVALID_TOKEN));
@@ -94,7 +57,7 @@ public class QueueService {
                             .message("입장이 가능합니다. 입찰 페이지로 이동합니다.")
                             .build());
                 })
-                .switchIfEmpty(redisRepository.getRank(waitingKey, userId)
+                .switchIfEmpty(trafficRepository.getRank(waitingKey, userId)
                         .map(rank -> QueueStatusResponse.builder()
                                 .auctionId(auctionId)
                                 .userId(userId)
@@ -113,7 +76,7 @@ public class QueueService {
     public Mono<String> cancelWaiting(UUID userId, UUID auctionId) {
         String waitingKey = "queue:auction:" + auctionId + ":waiting";
 
-        return redisRepository.removeWaitingUser(waitingKey, userId)
+        return trafficRepository.removeWaitingUser(waitingKey, userId)
                 .flatMap(removed -> {
                     if(removed <= 0)
                         return Mono.error(() -> new QueueException(QueueErrorCode.WAITING_USER_NOT_FOUND));
