+++ backend/queue-service/src/main/java/com/bidket/queue/infrastructure/schedule/QueueScheduler.java
@@ -1,6 +1,7 @@
 package com.bidket.queue.infrastructure.schedule;
 
-import com.bidket.queue.domain.repository.RedisRepository;
+import com.bidket.queue.domain.repository.QueueManagementRepository;
+import com.bidket.queue.domain.repository.QueueTrafficRepository;
 import com.bidket.queue.global.util.jwt.TokenProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,13 +18,14 @@ import java.util.UUID;
 @Component
 @RequiredArgsConstructor
 public class QueueScheduler {
-    private final RedisRepository redisRepository;
+    private final QueueManagementRepository managementRepository;
+    private final QueueTrafficRepository trafficRepository;
     private final TokenProvider tokenProvider;
 
     @Scheduled(fixedDelay = 1000)
     public void entranceSchedule() {
         log.info("스케줄링 시작");
-        redisRepository.getAllActiveAuctions()
+        managementRepository.getAllActiveAuctions()
                 .parallel()
                 .runOn(Schedulers.boundedElastic())
                 .flatMap(this::processAuction)
@@ -38,9 +40,9 @@ public class QueueScheduler {
         String activeKey = "queue:auction:" + auctionId + ":active";
         String waitingKey = "queue:auction:" + auctionId + ":waiting";
 
-        return redisRepository.getConfig(configKey)
+        return managementRepository.getConfig(configKey)
                 .flatMap(config ->
-                        redisRepository.activeUserCount(activeKey)
+                        trafficRepository.getActiveUserCount(activeKey)
                                 .flatMap(currentActive -> {
                                     long maxUser = config.getMaxActive();
                                     long availableSlots = maxUser - currentActive;
@@ -49,7 +51,7 @@ public class QueueScheduler {
                                     if (limit <= 0)
                                         return Mono.just(0L);
 
-                                    return redisRepository.popUserIdWaitingQueue(waitingKey, limit)
+                                    return trafficRepository.popUserIdWaitingQueue(waitingKey, limit)
                                             .flatMap(userIds -> {
                                                 if (userIds.isEmpty())
                                                     return Mono.just(0L);
@@ -61,10 +63,10 @@ public class QueueScheduler {
                                                 });
 
                                                 log.info("경매[{}] {} 명 입장", auctionId, userIds.size());
-                                                return redisRepository.addAllActiveUser(activeKey, userIds)
+                                                return trafficRepository.addAllActiveUser(activeKey, userIds)
                                                         .flatMap(added -> {
                                                             String tokenKey = "queue:token:" + auctionId;
-                                                            return redisRepository.saveToken(tokenKey, userTokens);
+                                                            return trafficRepository.saveToken(tokenKey, userTokens);
                                                         });
                                             });
 
@@ -74,7 +76,7 @@ public class QueueScheduler {
                 .switchIfEmpty(Mono.defer(() -> {
                     log.warn("경매[{}] 설정 만료, 관리 목록에서 제거", auctionId);
 
-                    return redisRepository.removeActiveAuction(auctionId)
+                    return managementRepository.removeActiveAuction(auctionId)
                             .then(Mono.empty());
                 }))
                 .then();
